+++ NEWS.md
@@ -5,6 +5,8 @@ Notable changes include:
  * deprecation of `slice()` in favor of the new, faster, more robust implementation in `dice()` 
  * complete overhaul of `slab()`, with new arguments, faster back-end, and weighted aggregation implemented (finally)
  * deprecation of `profile_compare()` in favor of the `NCSP()`--a complete overhaul based on Maynard et al., 2020
+   - site level attributes now handled by `compareSites()`
+   - variable weights now possible via argument
  * `perturb()` and `estimatePSCS()` are now vectorized, and optimized for larger collections 
 
 Incremental changes, should have no effect on previous code: